+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Neste arquivo (CHANGELOG.MD) você encontrará somente as mudanças referentes a
 ### Sobre os "BREAKING CHANGES"
 Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de serem pequenas, podem alterar o comportamento da funcionalidade caso não seja feita uma atualização, **preste muita atenção** nas breaking changes dentro das versões quando existirem.
 
+## [Não publicado]
+### Corrigido
+- `QasBtnDropdown`: corrigido dropdown para o uso do botão no menu para troca dos módulos.
+
 ## [3.15.0-beta.9] - 04-04-2024
 ## BREAKING CHANGES
 - `QasBtnDropdown`: